MAPPING: func_asw_fade and prop_asw_fade
Same as func_brush and prop_dynamic, but they will fade to a specified
alpha if the camera is third person and the marine is below their
origin.

Caveat: When a marine is above their origin, they will be fully opaque
after they stop fading. I didn't save the original render alpha
anywhere, but it would be simple to modify it to keep the original alpha
if mappers would actually need that functionality.